fix(golang): fix wrong file name (#48)

+++ src/golang/build/golang.sh
@@ -2,18 +2,38 @@
 
 set -e
 
+check_semver ${GOLANG_VERSION}
+
+
+if [[ ! "${MAJOR}" || ! "${MINOR}" || ! "${PATCH}" ]]; then
+  echo Invalid version: ${GOLANG_VERSION}
+  exit 1
+fi
+
+if [[ -d "/usr/local/go/${GOLANG_VERSION}" ]]; then
+  echo "Skipping, already installed"
+  exit 0
+fi
+
+# fix version
+GOLANG_FILE_VERSION=${GOLANG_VERSION}
+if [[ "${PATCH}" == "0" ]]; then
+  GOLANG_FILE_VERSION="${MAJOR}.${MINOR}"
+fi
+
 # go suggests: git svn bzr mercurial
 apt_install bzr mercurial
 
-curl -sSL https://dl.google.com/go/go${GOLANG_VERSION}.linux-amd64.tar.gz --output go.tgz
-tar -C /usr/local -xzf go.tgz
+mkdir -p /usr/local/go/${GOLANG_VERSION}
+curl -sSL https://dl.google.com/go/go${GOLANG_FILE_VERSION}.linux-amd64.tar.gz --output go.tgz
+tar --strip 1 -C /usr/local/go/${GOLANG_VERSION} -xzf go.tgz
 rm go.tgz
 
 export_env GOPATH "/go"
 export_env CGO_ENABLED 0
 export_env GOPROXY direct
 export_env GOSUMDB off
-export_path "/usr/local/go/bin:\$GOPATH/bin"
+export_path "/usr/local/go/${GOLANG_VERSION}/bin:\$GOPATH/bin"
 
 mkdir -p "$GOPATH/src" "$GOPATH/bin" "$GOPATH/pkg"
